Replaced ActionButtons with NewActionButtons
